requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.15.2
keras==2.0.8
opencv-python
h5py
imgaug
IPython[all]
pyqt5